Major commit.
Added the match_updater service and the db_worker service
Added a redis db for the match_updater and a postgres db for the db_worker services

Fixed minor bugs in the other services

A django project is currently included in the db_worker folder but not used. Will be removed later.

Api is ready to be tested, local tests failed probably due to internet issues.